fix:quotes & ui, resolve review errors

+++ routes/web.php
@@ -24,14 +24,16 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::get('/', [LandingController::class, 'index'])->name('home');
-Route::get('/movie/{movie:slug}', [MoviesController::class, 'index'])->name('bymovie');
-Route::get('/author/{author:username}', [UsersController::class, 'index'])->name('byauthor');
+
+Route::get('/movie/{movie:slug}', [MoviesController::class, 'index'])->name('movie.index');
+
+Route::get('/author/{author:username}', [UsersController::class, 'index'])->name('author.index');
 //language
 
-Route::get('/change-language/{locale}',[LanguageController::class, 'change'])->name('change.language');
+Route::get('/change-language/{locale}',[LanguageController::class, 'change'])->name('language.change');
 
 //authorisation
-Route::get('/login', function (){return view('create');})->name('auth.loginPage')->middleware('guest');
+Route::get('/login', function (){return view('login');})->name('auth.loginPage')->middleware('guest');
 
 Route::post('login', [AuthController::class, 'login'])->name('auth.login')->middleware('guest');
 Route::get('logout', [AuthController::class, 'logout'])->name('auth.logout')->middleware('auth');
@@ -41,20 +43,19 @@ Route::get('logout', [AuthController::class, 'logout'])->name('auth.logout')->mi
 Route::middleware(['auth'])->group(function () {
 Route::controller(MoviesController::class)->group(function (){
     Route::get('/admin',  'admin')->name('admin');
-    Route::get('/admin/show/{movie:slug}', 'show')->name('show-movie');
+
+    Route::get('/admin/show/{movie:slug}', 'show')->name('movie.show');
 
     Route::get('/admin/create/movie',  function (){
-                return view('admin.create-movie');})->name('create-movie');
-    Route::post('/admin/create/movie',  'storeMovie')->name('store-movie');
+                return view('admin.create-movie');})->name('movie.create');
+    Route::post('/admin/create/movie',  'store')->name('movie.store');
 
-    Route::get('/admin/edit/{movie:slug}', 'edit')->name('edit-movie');
-    Route::patch('/admin/edit/{movie:slug}',  'update')->name('update-movie');
+    Route::get('/admin/edit/{movie:slug}', 'edit')->name('movie.edit');
+    Route::patch('/admin/edit/{movie:slug}',  'update')->name('movie.update');
 
-    Route::get('/admin/{movie:slug}', 'view')->name('viewmovie');
-    Route::delete('/admin/movies/delete/{movie:slug}',  function (Movie $movie){
-        $movie->delete();
-        return back();
-    })->name('delete-movie');
+
+    Route::delete('/admin/movies/delete/{movie:slug}', 'delete'
+    )->name('movie.delete');
 });
 
 Route::controller(QuotesController::class)->group(function (){
@@ -65,11 +66,11 @@ Route::controller(QuotesController::class)->group(function (){
                     ]
                 );})->name('create-quote');
 
-    Route::post('/admin/show/{movie:slug}/quote',  'store')->name('store-quote');
+    Route::post('/admin/show/{movie:slug}/quote',  'store')->name('quote.store');
 
-    Route::get('/admin/show/{movie:slug}/{quote}/edit', 'edit')->name('edit-quote');
-    Route::patch('/admin/show/{movie:slug}/{quote}/edit',  'update')->name('update-quote');
+    Route::get('/admin/show/{movie:slug}/{quote}/edit', 'edit')->name('quote.edit');
+    Route::patch('/admin/show/{movie:slug}/{quote}/edit',  'update')->name('quote.update');
 
-    Route::post('/admin/show/{movie:slug}/{quote}/delete',  'view')->name('delete-quote');
+    Route::delete('/admin/show/{movie:slug}/{quote}/delete',  'delete')->name('quote.delete');
 });
 });
